Movement of Se atoms in upper layer
The electron irradiation on the model was simulated using the PKA technique, Random (PKA) atoms selected in the collision step. The energy transferred to the atoms from the collision with electrons during the irradiation simulation depends on the energy of the electron beam mass of the atom and scattering angle. This was done in accordance to the relations of the RUTHERFORD SCATTERING MODEL which was implemented in the molecular dynamics simulation.